add jenkins instance package
