+++ src/org/dsty/jenkins/package-info.groovy
@@ -1,0 +1,4 @@
+/**
+ * Get information about the Jenkins Instance.
+ */
+ package org.dsty.jenkins
