+++ tests/test_bash/test_bashClient.py
@@ -1,40 +1,3 @@
-import os
-from pathlib import Path
-from typing import ByteString
-
-import docker
-import pytest
-
-@pytest.fixture(scope='session')
-def client():
-    return docker.from_env()
-
-@pytest.fixture(scope='session')
-def container(client: docker.DockerClient):
-
-    image = os.getenv('RUNNER_IMAGE', 'iorunner')
-
-    jobs_path: str = str(Path(__file__, '../../../jobs').resolve())
-    lib_path: str = str(Path(__file__, '../../../').resolve())
-    
-    def create_container(job_name):
-
-        volumes = {
-            jobs_path: {
-                'bind': '/workspace',
-                'mode': 'ro'
-            },
-            lib_path: {
-                'bind': '/var/jenkins_home/pipeline-library',
-                'mode': 'ro'
-            }
-        }
-
-        return client.containers.run(image, f'-f /workspace/{job_name}', remove=True, volumes=volumes)
-
-
-    return create_container
-
 def test_call(container):
 
     job_output = container('bash/call')
