+++ tests/test_jenkins/test_Instance.py
@@ -1,0 +1,3 @@
+def test_plugins(container):
+
+    container('jenkins/plugins')
